# bdtheque_swagger_node/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Gestion d'une BDthèque
# during dev, should point to your local machine
host: localhost:8080
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /authors:
    # Configuration du chemin du contrôleur
    x-swagger-router-controller: authors
    # Précision du verbe HTTP 
    get:
      description: "Renvoie la liste complète des auteurs."
      # Précision sur la méthode
      operationId: authors
      responses:
        "200":
          description: "Renvoie tous les auteurs présent dans la base."
          schema: {
             type: "array",
             items: {
               $ref: "#/definitions/Authors"
             }
           }
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /authors/{id}:
    x-swagger-router-controller: authors
    get:
      description: "Renvoie un auteur suivant son identifiant technique."
      operationId: author
      parameters:
        - name: id
          in: path
          type: integer
          description: "l'identifiant technique de l'auteur"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Authors"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  Authors:
    properties:
      id:
        type: "number"
        description: "l'identifiant technique"
      last_name:
        type: "string"
        description: "le nom de famille"
      first_name:
        type: "string"
        description: "le prénom"
      date_of_birth:
        type: "string"
        description: "la date de naissance"
      place_of_birth:
        type: "string"
        description: "le lieux de naissance"
      nationality:
        type: "string"
        description: "la nationalité"
      series:
        type: "array"     
        description: "l'ensemble des séries de l'auteur"
        items: 
          type: "object"
          properties: 
            name: 
              type: "string"
              description: "Le nom de la série"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string